复习第一天
8-11：00

+++ src/abc往期学习2021年10月2日/aaa001初期学习集合/aaa021输入数据用方法判断大小.java
@@ -13,9 +13,14 @@ public class aaa021输入数据用方法判断大小 {
         Scanner xx = new Scanner(System.in);
         Random zz = new Random();
         System.out.println("请输入比较值：");
-        double x1 = xx.nextInt();
-        double z1 = zz.nextInt(1000);
-
+        // double x1 = xx.nextInt();
+        int x1 = xx.nextInt();
+        // double z1 = zz.nextInt(1000);
+        int z1 = zz.nextInt(1000);
+        //上面设置对比数据为int类型，但是最终展示的是double类型的mm
+        //所以结果带上小数点
+        // 效果为：两个int类型的数据比较后，得到double类型的数据结果
+        
         if (x1 > z1) {
             mm = x1;
             System.out.println("较大值为：" + mm);
